Share PartialAssistantMessage type between agent-service and main app
Extract the duplicated PartialAssistantMessage type into a shared/
directory that both the agent-service and main Next.js app import from.
This eliminates the risk of silent protocol divergence between the two
codebases.

Fixes #212
