+++ agent-service/src/stream-accumulator.ts
@@ -11,6 +11,9 @@
  * the frontend replaces the partial with the final version.
  */
 
+export type { PartialAssistantMessage } from '@shared/agent-types.js';
+import type { PartialAssistantMessage } from '@shared/agent-types.js';
+
 /**
  * Represents a content block being accumulated from stream deltas.
  */
@@ -22,29 +25,6 @@ interface AccumulatingContentBlock {
   id?: string;
   name?: string;
   input?: string; // accumulated JSON string, parsed at emission time
-}
-
-/**
- * A partial assistant message built from accumulated stream events.
- * Shaped to match the assistant message content structure so the frontend
- * can render it identically to a complete message.
- */
-export interface PartialAssistantMessage {
-  type: 'assistant';
-  /** Whether this is a partial (in-progress) message */
-  partial: true;
-  message: {
-    role: 'assistant';
-    content: Array<
-      | { type: 'text'; text: string }
-      | { type: 'tool_use'; id: string; name: string; input: Record<string, unknown> }
-    >;
-    model?: string;
-    stop_reason?: string | null;
-  };
-  parent_tool_use_id: string | null;
-  uuid: string;
-  session_id: string;
 }
 
 /**
