+++ shared/agent-types.ts
@@ -1,0 +1,32 @@
+/**
+ * Shared types for the SSE protocol between the agent service and the Next.js server.
+ *
+ * These types define the wire format for messages streamed from the agent service
+ * running inside containers to the Next.js application.
+ */
+
+/**
+ * A partial assistant message built from accumulated stream events.
+ * Shaped to match the assistant message content structure so the frontend
+ * can render it identically to a complete message.
+ *
+ * Generated by StreamAccumulator in the agent service, serialized as JSON
+ * over SSE to the Next.js server.
+ */
+export interface PartialAssistantMessage {
+  type: 'assistant';
+  /** Whether this is a partial (in-progress) message */
+  partial: true;
+  message: {
+    role: 'assistant';
+    content: Array<
+      | { type: 'text'; text: string }
+      | { type: 'tool_use'; id: string; name: string; input: Record<string, unknown> }
+    >;
+    model?: string;
+    stop_reason?: string | null;
+  };
+  parent_tool_use_id: string | null;
+  uuid: string;
+  session_id: string;
+}
